Reorganise components for section fields
