+++ src/components/Screens/screenIndex.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { get } from 'lodash';
 import StageEditorScreen from './StageEditorScreen';
-import SkipLogicEditorScreen from './SkipLogicEditorScreen';
 import Codebook from './CodebookScreen';
 import TypeEditorScreen from './TypeEditorScreen';
 import VariableEditorScreen from './VariableEditorScreen';
@@ -12,7 +11,6 @@ const NotFound = () => (<div> Screen not found </div>);
 
 const SCREEN_INDEX = {
   stage: StageEditorScreen,
-  skip: SkipLogicEditorScreen,
   codebook: Codebook,
   type: TypeEditorScreen,
   variable: VariableEditorScreen,
