--- src/components/SkipLogicEditor/index.js
@@ -1,2 +1,0 @@
-export { default } from './SkipLogicEditor';
-export { formName } from './SkipLogicEditor';
