+++ src/components/sections/FilteredEdgeType.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import Section from '@components/sections/Section';
 import Row from '@components/sections/Row';
-import NetworkFilter from '@components/sections/NetworkFilter';
-import EdgeTypeFields from '@components/sections/EdgeTypeFields';
+import NetworkFilter from '@components/sections/fields/NetworkFilter';
+import EdgeTypeFields from '@components/sections/fields/EdgeTypeFields';
 
 const FilteredEdgeType = props => (
   <Section>
