+++ src/components/sections/NodeType.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Section from '@components/sections/Section';
-import NodeTypeFields from '@components/sections/NodeTypeFields';
+import NodeTypeFields from '@components/sections/fields/NodeTypeFields';
 
 const FilteredNodeType = props => (
   <Section>
