+++ src/components/sections/fields/NodeTypeFields/withDisplayVariableOptions.js
